requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
django-cors-headers==3.2.1
django-extensions==3.1.1
django-filter==2.2.0
django-widget-tweaks==1.4.5
Django==3.0.2
djangorestframework==3.11.0
idna==2.8
Markdown==3.1.1
Pillow>=7.0.0
pytz==2019.3
requests==2.22.0
urllib3==1.25.7